Add an easier way to execute cd ..
This commit adds a special command to my bash shell, which maps `..` to
`cd ..`. The cool thing about it is, you can add how much dots as you
want, for example `.....` maps to `cd ../../../..`.

The implementation of this is a little hacky, but you can use it even
if you don't use my whole set of dotfiles.